reorganize kindling files
