+++ src/main/kotlin/dev/ashli/fire/Main.kt
@@ -1,45 +1,14 @@
 package dev.ashli.fire
 
-import dev.ashli.fire.parser.emitter.KindlingValue
+import dev.ashli.fire.parser.emitter.*
 
 fun main(args: Array<String>) {
 
-    val toEmit = KindlingValue.List(arrayOf(
-        KindlingValue.Identifier("def"),
-        KindlingValue.Text("constants"),
-        KindlingValue.List(arrayOf(
-            KindlingValue.List(arrayOf(
-                KindlingValue.Identifier("set-var"),
-                KindlingValue.Text("="),
-                KindlingValue.List(arrayOf(
-                    KindlingValue.List(arrayOf(
-                        KindlingValue.Identifier("global"),
-                        KindlingValue.Text("firstJoin")
-                    )),
-                    KindlingValue.List(arrayOf(
-                        KindlingValue.Identifier("num"),
-                        KindlingValue.Number(1.0)
-                    ))
-                ))
-            )),
-            KindlingValue.List(arrayOf(
-                KindlingValue.Identifier("set-var"),
-                KindlingValue.Text("="),
-                KindlingValue.List(arrayOf(
-                    KindlingValue.List(arrayOf(
-                        KindlingValue.Identifier("global"),
-                        KindlingValue.Text("clickerItem")
-                    )),
-                    KindlingValue.List(arrayOf(
-                        KindlingValue.Identifier("item"),
-                        KindlingValue.EscapedText("""{id:"minecraft:feather",Count:1,tag:{display:{Name:\'{"italic":"false","color":"#FFFFAA","text":"Clicker"}\'}}}""")
-                    ))
-                ))
-            ))
-        ))
+    val toEmit = KindlingHeader.PlayerEvent("join", arrayOf(
+        KindlingAction.PlayerAction("SendMessage", "", arrayOf(KindlingArgument.Text("hi %default"))),
     ))
 
-    val output = KindlingValue.emitValue(toEmit)
+    val output = emitValue(emitHeader(toEmit))
     println(output)
 
     return
