+++ src/main/kotlin/dev/ashli/fire/parser/emitter/KindlingArgument.kt
@@ -2,104 +2,6 @@ package dev.ashli.fire.parser.emitter
 
 import kotlin.math.pow
 
-sealed class KindlingValue {
-    class Text(val value: String) : KindlingValue()
-    class EscapedText(val value: String) : KindlingValue()
-    class Number(val value: Double) : KindlingValue()
-    class List(var value: Array<KindlingValue>) : KindlingValue() {}
-    class Identifier(val value: String) : KindlingValue()
-}
-
-fun emitValue(value: KindlingValue): String {
-    return when(value) {
-        is KindlingValue.Number -> "${value.value}"
-        is KindlingValue.Text -> "\"${value.value}\""
-        is KindlingValue.EscapedText -> "'${value.value}'"
-        is KindlingValue.List -> {
-            var builder = "("
-            for(item in value.value) {
-                builder = "$builder ${emitValue(item)}"
-            }
-            builder = "$builder )"
-            builder
-        }
-        is KindlingValue.Identifier -> value.value
-    }
-}
-sealed class KindlingHeader {
-    class PlayerEvent(val name: String, val code: Array<KindlingAction>) : KindlingHeader()
-    class EntityEvent(val name: String, val code: Array<KindlingAction>) : KindlingHeader()
-    class Function(val name: String, val code: Array<KindlingAction>) : KindlingHeader()
-    class Process(val name: String, val code: Array<KindlingAction>) : KindlingHeader()
-}
-
-fun emitHeader(self: KindlingHeader): KindlingValue {
-    return when(self) {
-        is KindlingHeader.PlayerEvent -> {
-            KindlingValue.List(arrayOf(
-                KindlingValue.Identifier("player-event"),
-                KindlingValue.Text(self.name),
-                KindlingValue.List(emitActions(self.code))
-            ))
-        }
-        is KindlingHeader.EntityEvent -> {
-            KindlingValue.List(arrayOf(
-                KindlingValue.Identifier("entity-event"),
-                KindlingValue.Text(self.name),
-                KindlingValue.List(emitActions(self.code))
-            ))
-        }
-        is KindlingHeader.Function -> {
-            KindlingValue.List(arrayOf(
-                KindlingValue.Identifier("def"),
-                KindlingValue.Text(self.name),
-                KindlingValue.List(emitActions(self.code))
-            ))
-        }
-        is KindlingHeader.Process -> {
-            KindlingValue.List(arrayOf(
-                KindlingValue.Identifier("process"),
-                KindlingValue.Text(self.name),
-                KindlingValue.List(emitActions(self.code))
-            ))
-        }
-    }
-
-}
-
-sealed class KindlingAction {
-    class PlayerAction(val type: String, val selector: String, val arguments: Array<KindlingArgument>) : KindlingAction()
-    class EntityAction(val type: String, val selector: String, val arguments: Array<KindlingArgument>) : KindlingAction()
-}
-
-fun emitAction(self: KindlingAction): KindlingValue {
-    return when(self) {
-        is KindlingAction.PlayerAction -> {
-            KindlingValue.List(arrayOf(
-                KindlingValue.Identifier("player-action"),
-                KindlingValue.Text(self.type),
-                KindlingValue.Text(self.selector),
-                KindlingValue.List(self.arguments.map { x -> emitArgument(x) }.toTypedArray())
-            ))
-        }
-        is KindlingAction.EntityAction -> {
-            KindlingValue.List(arrayOf(
-                KindlingValue.Identifier("entity-action"),
-                KindlingValue.Text(self.type),
-                KindlingValue.Text(self.selector),
-                KindlingValue.List(self.arguments.map { x -> emitArgument(x) }.toTypedArray())
-            ))
-        }
-    }
-}
-
-fun emitActions(self: Array<KindlingAction>): Array<KindlingValue> {
-    var out: Array<KindlingValue> = arrayOf()
-    for(i in self) {
-        out += emitAction(i)
-    }
-    return out
-}
 
 sealed class KindlingArgument {
     class Text(val value: String) : KindlingArgument()
@@ -327,26 +229,4 @@ fun emitArgument(value: KindlingArgument): KindlingValue {
         }
 
     }
-}
-
-class ParticleSettings {
-    val amount: Int = 0
-    val spreadX: Int? = null
-    val spreadY: Int? = null
-    val motionX: Int? = null
-    val motionY: Int? = null
-    val motionZ: Int? = null
-    val roll: Int? = null
-    val size: Int? = null
-    val rgb: ParticleRGB? = null
-    val material: String? = null // should be a valid material
-    val variationColor: Int? = null // between 1-100
-    val variationMotion: Int? = null // between 1-100
-    val variationSize: Int? = null // between 1-100
-}
-
-class ParticleRGB {
-    val r: Int = 0
-    val g: Int = 0
-    val b: Int = 0
 }
